pyproject.toml
[tool.poetry]
name = "fastapi-task-manager"
version = "0.1.0"
description = "this is a task manager application"
authors = ["Pankaj"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
sqlalchemy = "^2.0.30"
databases = "^0.9.0"
asyncpg = "^0.29.0"
passlib = "^1.7.4"
bcrypt = "^4.1.3"
python-jose = "^3.3.0"
pydantic-settings = "^2.3.3"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
